backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.0.1
pytesseract==0.3.10
requests==2.32.4
